+++ src/ViewVolume.cpp
@@ -296,7 +296,7 @@ void ViewVolume::updateFrustum()
     mFrustum[0][3] = clip[15] - clip[12];
 
     /* Normalize the result */
-    t = sqrt(mFrustum[0][0] * mFrustum[0][0] +
+    t = sqrtf(mFrustum[0][0] * mFrustum[0][0] +
             mFrustum[0][1] * mFrustum[0][1] +
             mFrustum[0][2] * mFrustum[0][2]);
     mFrustum[0][0] /= t;
@@ -311,7 +311,7 @@ void ViewVolume::updateFrustum()
     mFrustum[1][3] = clip[15] + clip[12];
 
     /* Normalize the result */
-    t = sqrt(mFrustum[1][0] * mFrustum[1][0] +
+    t = sqrtf(mFrustum[1][0] * mFrustum[1][0] +
             mFrustum[1][1] * mFrustum[1][1] +
             mFrustum[1][2] * mFrustum[1][2]);
     mFrustum[1][0] /= t;
@@ -326,7 +326,7 @@ void ViewVolume::updateFrustum()
     mFrustum[2][3] = clip[15] + clip[13];
 
     /* Normalize the result */
-    t = sqrt(mFrustum[2][0] * mFrustum[2][0] +
+    t = sqrtf(mFrustum[2][0] * mFrustum[2][0] +
             mFrustum[2][1] * mFrustum[2][1] +
             mFrustum[2][2] * mFrustum[2][2]);
     mFrustum[2][0] /= t;
@@ -341,7 +341,7 @@ void ViewVolume::updateFrustum()
     mFrustum[3][3] = clip[15] - clip[13];
 
     /* Normalize the result */
-    t = sqrt(mFrustum[3][0] * mFrustum[3][0] +
+    t = sqrtf(mFrustum[3][0] * mFrustum[3][0] +
             mFrustum[3][1] * mFrustum[3][1] +
             mFrustum[3][2] * mFrustum[3][2]);
     mFrustum[3][0] /= t;
@@ -356,7 +356,7 @@ void ViewVolume::updateFrustum()
     mFrustum[4][3] = clip[15] - clip[14];
 
     /* Normalize the result */
-    t = sqrt(mFrustum[4][0] * mFrustum[4][0] +
+    t = sqrtf(mFrustum[4][0] * mFrustum[4][0] +
             mFrustum[4][1] * mFrustum[4][1] +
             mFrustum[4][2] * mFrustum[4][2]);
     mFrustum[4][0] /= t;
@@ -371,7 +371,7 @@ void ViewVolume::updateFrustum()
     mFrustum[5][3] = clip[15] + clip[14];
 
     /* Normalize the result */
-    t = sqrt(mFrustum[5][0] * mFrustum[5][0] +
+    t = sqrtf(mFrustum[5][0] * mFrustum[5][0] +
             mFrustum[5][1] * mFrustum[5][1] +
             mFrustum[5][2] * mFrustum[5][2]);
     mFrustum[5][0] /= t;
